+++ packages/terra-form-checkbox/tests/wdio/form-checkbox-spec.js
@@ -10,6 +10,7 @@ describe('Checkbox', () => {
     '--terra-form-checkbox-container-margin-top': '0',
     '--terra-form-checkbox-font-size': '1.286rem',
     '--terra-form-checkbox-line-height': '1.4',
+    '--terra-form-checkbox-padding-bottom': '0',
     '--terra-form-checkbox-padding-left': '1.2rem',
     '--terra-form-checkbox-native-input-clip': 'rect(0 0 0 0)',
     '--terra-form-checkbox-native-input-cursor': 'default',
@@ -20,7 +21,6 @@ describe('Checkbox', () => {
     '--terra-form-checkbox-native-input-overflow': 'hidden',
     '--terra-form-checkbox-native-input-padding': '0',
     '--terra-form-checkbox-native-input-width': '1px',
-    '--terra-form-checkbox-label-text-margin-bottom': '0',
     '--terra-form-checkbox-label-text-margin-left': '0.643rem',
     '--terra-form-checkbox-background-color': '#fff',
     '--terra-form-checkbox-background-image': 'none',
@@ -33,9 +33,9 @@ describe('Checkbox', () => {
   };
 
   const hoverThemingProperties = {
-    '--terra-form-checkbox-hover-border-color': '#77b1fd',
+    '--terra-form-checkbox-hover-border-color': '#800080',
     '--terra-form-checkbox-hover-border-width': '2px',
-    '--terra-form-checkbox-checked-hover-border-color': '#77b1fd',
+    '--terra-form-checkbox-checked-hover-border-color': '#800080',
     '--terra-form-checkbox-checked-hover-border-width': '2px',
   };
 
@@ -45,17 +45,17 @@ describe('Checkbox', () => {
   };
 
   const focusRingThemingProperties = {
-    '--terra-form-checkbox-focus-ring-background-color': 'orange',
-    '--terra-form-checkbox-focus-ring-height': '10px',
-    '--terra-form-checkbox-focus-ring-left': '16.3px',
-    '--terra-form-checkbox-focus-ring-top': '9.2px',
-    '--terra-form-checkbox-focus-ring-width': '10px',
+    '--terra-form-checkbox-focus-ring-background-color': 'rgba(128, 0, 128, 0.25)',
+    '--terra-form-checkbox-focus-ring-height': '32px',
+    '--terra-form-checkbox-focus-ring-left': '-8px',
+    '--terra-form-checkbox-focus-ring-top': '-3px',
+    '--terra-form-checkbox-focus-ring-width': '32px',
     '--terra-form-checkbox-focus-ring-border-radius': '50%',
   };
 
   const checkedThemingProperties = {
-    '--terra-form-checkbox-checked-background-color': '#3a67a4',
-    '--terra-form-checkbox-checked-border': '2px solid #3a67a4',
+    '--terra-form-checkbox-checked-background-color': '#f00',
+    '--terra-form-checkbox-checked-border': '2px solid #f00',
     '--terra-form-checkbox-checked-border-radius': '0.2714rem',
     '--terra-form-checkbox-checked-height': '1.286rem',
     '--terra-form-checkbox-checked-line-height': '0.9',
@@ -76,11 +76,6 @@ describe('Checkbox', () => {
     '--terra-form-checkbox-disabled-border-width': '1px',
     '--terra-form-checkbox-checked-disabled-border-width': '1px',
     '--terra-form-checkbox-disabled-symbol': 'url("data:image/svg+xml;charset=US-ASCII,%3Csvg%20version%3D%221.1%22%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20xmlns%3Axlink%3D%22http%3A%2F%2Fwww.w3.org%2F1999%2Fxlink%22%20viewBox%3D%220%200%2016%2016%22%20enable-background%3D%22new%200%200%2016%2016%22%20xml%3Aspace%3D%22preserve%22%3E%3Cpolygon%20fill%3D%22%23cfd0d1%22%20points%3D%225.6%2C9.6%202.4%2C6.4%200%2C8.8%205.6%2C14.4%2016%2C4%2013.6%2C1.6%20%22%3E%3C%2Fpolygon%3E%3C%2Fsvg%3E")',
-  };
-
-  const disabledHoverThemingProperties = {
-    '--terra-form-checkbox-disabled-hover-border-color': '#cfd0d1',
-    '--terra-form-checkbox-disabled-hover-border-width': '1px',
   };
 
   describe('Default', () => {
@@ -111,6 +106,7 @@ describe('Checkbox', () => {
       beforeEach(() => {
         browser.waitForVisible('#default');
         browser.click('[for="default"]');
+        browser.click('#site');
       });
 
       Terra.should.matchScreenshot();
@@ -131,12 +127,17 @@ describe('Checkbox', () => {
       Terra.should.beAccessible();
       Terra.should.themeCombinationOfCustomProperties({
         testName: 'themed',
-        properties: { ...themingProperties, ...focusThemingProperties },
+        properties: { ...themingProperties, ...checkedThemingProperties, ...focusThemingProperties },
       });
 
       Terra.should.themeCombinationOfCustomProperties({
         testName: 'custom - Focus Ring',
-        properties: { ...themingProperties, ...focusRingThemingProperties },
+        properties: {
+          ...themingProperties,
+          ...checkedThemingProperties,
+          ...focusThemingProperties,
+          ...focusRingThemingProperties,
+        },
       });
     });
   });
@@ -148,7 +149,7 @@ describe('Checkbox', () => {
     Terra.should.matchScreenshot();
     Terra.should.themeCombinationOfCustomProperties({
       testName: 'themed',
-      properties: { ...themingProperties, ...disabledThemingProperties },
+      properties: { ...themingProperties, ...checkedThemingProperties, ...disabledThemingProperties },
     });
 
     describe('Disabled Hover', () => {
@@ -161,7 +162,7 @@ describe('Checkbox', () => {
       Terra.should.matchScreenshot();
       Terra.should.themeCombinationOfCustomProperties({
         testName: 'themed',
-        properties: { ...themingProperties, ...disabledHoverThemingProperties },
+        properties: { ...themingProperties, ...checkedThemingProperties, ...disabledThemingProperties },
       });
     });
   });