+++ packages/terra-form-radio/tests/wdio/form-radio-spec.js
@@ -1,20 +1,17 @@
-/* eslint-disable no-unused-expressions */
-/* global browser, Terra, beforeEach */
+/* global before, browser, Terra */
 
 describe('Radio', () => {
-  beforeEach(() => {
-    browser.setViewportSize(Terra.viewports('tiny')[0]);
-  });
+  before(() => browser.setViewportSize(Terra.viewports('medium')[0]));
 
   const themingProperties = {
-    '--terra-form-radio-container-margin-right': '1rem',
+    '--terra-form-radio-font-color': '#111619',
+    '--terra-form-radio-container-margin-bottom': '0.625rem',
+    '--terra-form-radio-container-margin-right': '1.071rem',
     '--terra-form-radio-container-margin-top': '0',
-    '--terra-form-radio-font-size, 1rem': '1.1rem',
-    '--terra-form-radio-line-height': '1.7',
-    '--terra-form-radio-native-input-margin-left': '0.3rem',
-    '--terra-form-radio-native-input-margin-right': '0.3rem',
-    '--terra-form-radio-label-text-margin-bottom': '0',
-    '--terra-form-radio-label-text-margin-left': '1.5em',
+    '--terra-form-radio-font-size': '1.286rem',
+    '--terra-form-radio-line-height': '1.4',
+    '--terra-form-radio-padding-bottom': '0',
+    '--terra-form-radio-padding-left': '1.2rem',
     '--terra-form-radio-native-input-clip': 'rect(0 0 0 0)',
     '--terra-form-radio-native-input-cursor': 'default',
     '--terra-form-radio-native-input-font-size': 'none',
@@ -22,8 +19,9 @@ describe('Radio', () => {
     '--terra-form-radio-native-input-margin-left': 'none',
     '--terra-form-radio-native-input-margin-right': 'none',
     '--terra-form-radio-native-input-overflow': 'hidden',
-    '--terra-form-radio-native-input-padding': 0,
+    '--terra-form-radio-native-input-padding': '0',
     '--terra-form-radio-native-input-width': '1px',
+    '--terra-form-radio-label-text-margin-left': '0.643rem',
     '--terra-form-radio-outer-ring-background-color': 'rgb(255, 255, 255)',
     '--terra-form-radio-outer-ring-background-image': 'none',
     '--terra-form-radio-outer-ring-border': '#a2c2e0 solid 1px',
@@ -52,26 +50,26 @@ describe('Radio', () => {
   };
 
   const focusRingThemingProperties = {
-    '--terra-form-radio-focus-ring-background-color': 'rgba(74, 144, 226, 0.25)',
-    '--terra-form-radio-focus-ring-height': '30px',
-    '--terra-form-radio-focus-ring-left': '-5px',
-    '--terra-form-radio-focus-ring-top': '-5px',
-    '--terra-form-radio-focus-ring-width': '30px',
+    '--terra-form-radio-focus-ring-background-color': 'rgba(128, 0, 128, 0.25)',
+    '--terra-form-radio-focus-ring-height': '32px',
+    '--terra-form-radio-focus-ring-left': '-3px',
+    '--terra-form-radio-focus-ring-top': '-3px',
+    '--terra-form-radio-focus-ring-width': '32px',
   };
 
   const checkedThemingProperties = {
     '--terra-form-radio-outer-ring-checked-background-color': '#fff',
     '--terra-form-radio-outer-ring-checked-border': '#4a90e2 solid 1px',
+    '--terra-form-radio-outer-ring-checked-background-image': 'none',
     '--terra-form-radio-outer-ring-checked-border-radius': '50%',
     '--terra-form-radio-outer-ring-checked-height': '20px',
     '--terra-form-radio-outer-ring-checked-margin-top': '4px',
     '--terra-form-radio-outer-ring-checked-width': '20px',
-    '--terra-form-radio-inner-ring-checked-background-color': 'rgb(255, 0, 0)',
-    '--terra-form-radio-outer-ring-checked-background-image': 'none',
+    '--terra-form-radio-inner-ring-checked-background-color': '#f00',
     '--terra-form-radio-inner-ring-checked-border': 'rgb(255, 0, 0) solid 1px',
     '--terra-form-radio-inner-ring-checked-box-shadow': '0 1px 1px 0 #a43531 inset',
     '--terra-form-radio-inner-ring-checked-height': '14px',
-    '--terra-form-radio-inner-ring-checked-transition': 'all 0.3s cubic-bezier(0.78, 0.14, 0.15, 0.86)',
+    '--terra-form-radio-inner-ring-checked-transition': 'all 0.2s ease all',
     '--terra-form-radio-inner-ring-checked-width': '14px',
   };
 
@@ -95,9 +93,7 @@ describe('Radio', () => {
 
 
   describe('Default', () => {
-    beforeEach(() => {
-      browser.url('/#/raw/tests/terra-form-radio/form-radio/radio/default-radio');
-    });
+    beforeEach(() => browser.url('/#/raw/tests/terra-form-radio/form-radio/radio/default-radio'));
 
     Terra.should.matchScreenshot();
     Terra.should.beAccessible();
@@ -124,6 +120,7 @@ describe('Radio', () => {
       beforeEach(() => {
         browser.waitForVisible('#default');
         browser.click('[for="default"]');
+        browser.click('#site');
       });
 
       Terra.should.matchScreenshot();
@@ -144,12 +141,17 @@ describe('Radio', () => {
       Terra.should.beAccessible();
       Terra.should.themeCombinationOfCustomProperties({
         testName: 'themed',
-        properties: { ...themingProperties, ...focusThemingProperties },
+        properties: { ...themingProperties, ...checkedThemingProperties, ...focusThemingProperties },
       });
 
       Terra.should.themeCombinationOfCustomProperties({
         testName: 'custom - Focus Ring',
-        properties: { ...themingProperties, ...focusRingThemingProperties },
+        properties: {
+          ...themingProperties,
+          ...checkedThemingProperties,
+          ...focusThemingProperties,
+          ...focusRingThemingProperties,
+        },
       });
     });
   });
@@ -163,7 +165,7 @@ describe('Radio', () => {
     Terra.should.beAccessible();
     Terra.should.themeCombinationOfCustomProperties({
       testName: 'themed',
-      properties: { ...themingProperties, ...disabledThemingProperties },
+      properties: { ...themingProperties, ...checkedThemingProperties, ...disabledThemingProperties },
     });
   });
 
@@ -174,7 +176,7 @@ describe('Radio', () => {
     Terra.should.beAccessible();
     Terra.should.themeCombinationOfCustomProperties({
       testName: 'themed',
-      properties: { ...themingProperties, ...labelIsHiddenThemingProperties },
+      properties: { ...themingProperties, ...checkedThemingProperties, ...labelIsHiddenThemingProperties },
     });
   });
 });